+++ packages/sdk/CHANGELOG.md
@@ -11,7 +11,7 @@ commit that affects the code in a way that consumers might care about should inc
 
 ## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.33...HEAD)
 
-No changes.
+-   Fix a regression where `UI.ConfirmationDialog` would crash the block.
 
 ## [0.0.33](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.32...@airtable/blocks@0.0.33) - 2019-09-26